requirements.txt
Django==1.6.5
South==0.8.4
boto==2.29.1
dj-database-url==0.3.0
dj-static==0.0.5
django-braces==1.4.0
django-filter==0.7
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.3.13
drf-extensions==0.2.3
gunicorn==18.0
psycopg2==2.5.3
pystache==0.5.4
six==1.6.1
static==1.0.2
wsgiref==0.1.2
